+++ src/booking.js
@@ -1,6 +1,6 @@
 /*******************************
 *  Recras integration library  *
-*  v 2.0.3                     *
+*  v 2.1.0-alpha               *
 *******************************/
 
 class RecrasBooking {
@@ -363,7 +363,7 @@ class RecrasBooking {
         } else if (bookingSize > bsMaximum) {
             this.setMinMaxAmountWarning('bookingsize', bsMaximum, 'maximum');
         }
-        this.maybeDisableBookButton();
+        this.maybeShowInlineErrors();
     }
 
     checkDependencies() {
@@ -407,7 +407,7 @@ class RecrasBooking {
             }
         });
 
-        this.maybeDisableBookButton();
+        this.maybeShowInlineErrors();
     }
 
     checkDiscountAndVoucher() {
@@ -488,6 +488,55 @@ class RecrasBooking {
                 }
             });
         });
+    }
+
+    contactFormIsEmpty() {
+        let isEmpty = true;
+        let els = this.findElements('.recras-contactform input, .recras-contactform select, .recras-contactform textarea');
+        let formValues = [...els].map(el => el.value);
+        for (let val of formValues) {
+            if (val !== '') {
+                isEmpty = false;
+            }
+        }
+        return isEmpty;
+    }
+
+    contactFormInvalidFields() {
+        let invalid = [];
+        let required = this.contactFormRequiredFields();
+
+        let els = this.findElements('.recras-contactform :invalid');
+        for (let el of els) {
+            if (!required.includes(el) && el.type !== 'radio') {
+                invalid.push(el);
+            }
+        }
+        return invalid;
+    }
+
+    contactFormRequiredFields() {
+        let isEmpty = [];
+        let els = this.findElements('.recras-contactform :required');
+        for (let el of els) {
+            if (el.type === 'radio') {
+                const checkedRadio = this.findElement('.recras-contactform [name="' + el.name + '"]:checked');
+                if (!checkedRadio) {
+                    const radioEls = [...this.findElements('.recras-contactform [name="' + el.name + '"]')];
+                    const radioElLast = radioEls.at(-1);
+                    if (!isEmpty.includes(radioElLast)) {
+                        isEmpty.push(radioElLast);
+                    }
+                }
+            } else if (el.value === undefined || el.value === '') {
+                isEmpty.push(el);
+            }
+        }
+        return isEmpty;
+    }
+
+    contactFormRequiredIsEmpty() {
+        return this.contactFormRequiredFields().length > 0;
     }
 
     contactFormValid() {
@@ -762,41 +811,54 @@ class RecrasBooking {
         }
     }
 
-    maybeDisableBookButton() {
-        let button = this.findElement('.bookPackage');
-        if (!button) {
-            return false;
-        }
+    bookingErrors() {
+        let bookingDisabledReasons = {};
 
-        let bookingDisabledReasons = [];
-        if (this.requiresProduct) {
-            bookingDisabledReasons.push('BOOKING_DISABLED_REQUIRED_PRODUCT');
+        if (!this.hasAtLeastOneProduct(this.selectedPackage)) {
+            bookingDisabledReasons.amountsInvalid = 'BOOKING_DISABLED_NO_PRODUCTS';
+            return bookingDisabledReasons;
         }
         if (!this.amountsValid(this.selectedPackage)) {
-            bookingDisabledReasons.push('BOOKING_DISABLED_AMOUNTS_INVALID');
+            bookingDisabledReasons.amountsInvalid = 'BOOKING_DISABLED_AMOUNTS_INVALID';
+            return bookingDisabledReasons;
         }
+        if (this.requiresProduct) {
+            bookingDisabledReasons.requiresProduct = 'BOOKING_DISABLED_REQUIRED_PRODUCT';
+        }
+
         if (!this.selectedDate) {
-            bookingDisabledReasons.push('BOOKING_DISABLED_INVALID_DATE');
+            bookingDisabledReasons.dateInvalid = 'BOOKING_DISABLED_INVALID_DATE';
+            return bookingDisabledReasons;
         }
         if (!this.selectedTime) {
-            bookingDisabledReasons.push('BOOKING_DISABLED_INVALID_TIME');
+            bookingDisabledReasons.timeInvalid = 'BOOKING_DISABLED_INVALID_TIME';
+            return bookingDisabledReasons;
         }
-        if (!this.contactFormValid()) {
-            bookingDisabledReasons.push('BOOKING_DISABLED_CONTACT_FORM_INVALID');
+
+        if (this.contactFormIsEmpty()) {
+            bookingDisabledReasons.contactFormInvalid = 'BOOKING_DISABLED_CONTACT_FORM_EMPTY';
+        } else if (this.contactFormRequiredIsEmpty() || !this.contactFormValid()) {
+            // Special case
+            bookingDisabledReasons.contactFormRequired = true;
+            return bookingDisabledReasons;
         }
 
         const agreeEl = this.findElement('#recrasAgreeToAttachments');
         if (agreeEl && !agreeEl.checked) {
-            bookingDisabledReasons.push('BOOKING_DISABLED_AGREEMENT');
+            bookingDisabledReasons.notAgreed = 'BOOKING_DISABLED_AGREEMENT';
         }
 
-        if (bookingDisabledReasons.length > 0) {
-            const reasonsList = bookingDisabledReasons.map(reason => this.languageHelper.translate(reason)).join('<li>');
-            this.findElement('#bookingErrors').innerHTML = `<ul><li>${ reasonsList }</ul>`;
-            button.setAttribute('disabled', 'disabled');
-        } else {
-            this.findElement('#bookingErrors').innerHTML = '';
-            button.removeAttribute('disabled');
+        return bookingDisabledReasons;
+    }
+
+    bookingHasErrors() {
+        let bookingDisabledReasons = this.bookingErrors();
+        return Object.keys(bookingDisabledReasons).length > 0;
+    }
+
+    maybeShowInlineErrors() {
+        if (this.bookingHasErrors()) {
+            this.showInlineErrors();
         }
     }
 
@@ -895,6 +957,11 @@ class RecrasBooking {
         return counts;
     }
 
+    removeErrors(parentQS = '') {
+        [...this.findElements(parentQS + '.booking-error')].forEach(el => {
+            el.parentNode.removeChild(el);
+        });
+    }
     removeWarnings() {
         [...this.findElements('.booking-error:not(#bookingErrors)')].forEach(el => {
             el.parentNode.removeChild(el);
@@ -983,7 +1050,10 @@ class RecrasBooking {
         this.findElement('.standard-attachments').innerHTML = attachmentHtml;
         const agreeEl = this.findElement('#recrasAgreeToAttachments');
         if (agreeEl) {
-            agreeEl.addEventListener('change', this.maybeDisableBookButton.bind(this));
+            agreeEl.addEventListener('change', () => {
+                this.removeErrors('.standard-attachments ');
+                this.maybeShowInlineErrors();
+            });
         }
     }
 
@@ -1070,7 +1140,6 @@ class RecrasBooking {
             </div>`;
             this.appendHtml(html);
             this.findElement('.bookPackage').addEventListener('click', this.submitBooking.bind(this));
-            this.maybeDisableBookButton();
             this.updateProductAmounts();
         });
     }
@@ -1111,8 +1180,9 @@ class RecrasBooking {
                 this.eventHelper.sendEvent(RecrasEventHelper.PREFIX_BOOKING, RecrasEventHelper.EVENT_BOOKING_CONTACT_FORM_SHOWN);
 
                 [...this.findElements('[name^="contactformulier"]')].forEach(el => {
-                    el.addEventListener('input', this.maybeDisableBookButton.bind(this));
                     el.addEventListener('input', () => {
+                        this.removeErrors('.recras-contactform ');
+                        this.maybeShowInlineErrors();
                         if (this.contactFormValid()) {
                             this.nextSectionActive('.recras-contactform', '.recras-finalise');
                         }
@@ -1162,7 +1232,7 @@ class RecrasBooking {
         }
 
         this.findElement('#discountcode').removeAttribute('disabled');
-        this.maybeDisableBookButton();
+        this.maybeShowInlineErrors();
     }
 
     calendarOnDraw(pika, packageId) {
@@ -1206,7 +1276,7 @@ class RecrasBooking {
     selectTime(time) {
         this.eventHelper.sendEvent(RecrasEventHelper.PREFIX_BOOKING, RecrasEventHelper.EVENT_BOOKING_TIME_SELECTED);
         this.selectedTime = time;
-        this.maybeDisableBookButton();
+        this.maybeShowInlineErrors();
     }
 
     showDateTimeSelection(pack) {
@@ -1255,7 +1325,8 @@ class RecrasBooking {
                 this.previewTimes();
             }
 
-            this.maybeDisableBookButton();
+            this.removeErrors('.recras-datetime ');
+            this.maybeShowInlineErrors();
         });
     }
 
@@ -1370,6 +1441,77 @@ ${ msgs[1] }</p></div>`);
         return attachments;
     }
 
+    contactformErrors() {
+        this.removeErrors('.recras-contactform ');
+        for (let el of this.contactFormRequiredFields()) {
+            let labelEl = el.parentNode.closest('.recras-contactform-item').querySelector('label');
+            const requiredText = this.languageHelper.translate('CONTACT_FORM_FIELD_REQUIRED', { FIELD_NAME: labelEl.innerText });
+            el.parentNode.insertAdjacentHTML(
+                'afterend',
+                `<div class="booking-error">${ requiredText }</div>`
+            );
+        }
+        for (let el of this.contactFormInvalidFields()) {
+            let labelEl = el.parentNode.closest('.recras-contactform-item').querySelector('label');
+            const invalidText = this.languageHelper.translate('CONTACT_FORM_FIELD_INVALID', { FIELD_NAME: labelEl.innerText });
+            el.parentNode.insertAdjacentHTML(
+                'afterend',
+                `<div class="booking-error">${ invalidText }</div>`
+            );
+        }
+    }
+
+    errorAtPosition(key, msg) {
+        let pos;
+        let qs;
+
+        switch (key) {
+            case 'requiresProduct':
+            case 'amountsInvalid':
+                pos = 'beforeend';
+                qs = '.recras-amountsform';
+                break;
+            case 'dateInvalid':
+                pos = 'afterend';
+                qs = '.recras-onlinebooking-date';
+                break;
+            case 'timeInvalid':
+                pos = 'beforeend';
+                qs = '.recras-datetime';
+                break;
+            case 'contactFormInvalid':
+                pos = 'beforeend';
+                qs = '.recras-contactform';
+                break;
+            case 'notAgreed':
+                pos = 'beforeend';
+                qs = '.standard-attachments';
+                break;
+            case 'contactFormRequired':
+                this.contactformErrors();
+                break;
+            default:
+                pos = 'beforeend';
+                qs = '.bookPackage';
+                break;
+        }
+        const el = this.findElement(qs);
+        if (el) {
+            el.insertAdjacentHTML(
+                pos,
+                `<p class="booking-error">${ this.languageHelper.translate(msg) }</p>`
+            );
+        }
+    }
+
+    showInlineErrors() {
+        this.removeErrors();
+        let errorMsgs = this.bookingErrors();
+        for (const [key, msg] of Object.entries(errorMsgs)) {
+            this.errorAtPosition(key, msg);
+        }
+    }
+
     formatGA4Items() {
         let items = [];
 
@@ -1408,6 +1550,10 @@ ${ msgs[1] }</p></div>`);
     }
 
     submitBooking() {
+        if (this.bookingHasErrors()) {
+            this.showInlineErrors();
+            return false;
+        }
         let productCounts = this.productCounts().map(line => line.aantal);
         let productSum = productCounts.reduce((a, b) => a + b, 0);
         if (this.bookingSize() === 0 && productSum === 0) {
@@ -1525,6 +1671,7 @@ ${ msgs[1] }</p></div>`);
         this.loadingIndicatorHide();
         this.availableDays = [];
 
+        this.removeErrors();
         this.removeWarnings();
         this.checkDependencies();
         this.checkMinMaxAmounts();