+++ test/js-unit/recrasBookingSpec.js
@@ -368,9 +368,11 @@ describe('RecrasBooking', () => {
                     resolve();
                 });
             });
+            this.rb.bookingHasErrors = () => false;
+            this.rb.findElement = () => document.createElement('div');
 
             this.rb.bookingSize = () => 5;
-            this.rb.submitBooking();//
+            this.rb.submitBooking();
             expect(this.rb.postJson).toHaveBeenCalledWith('onlineboeking/reserveer', jasmine.objectContaining({
                 betaalmethode: 'factuur',
             }));
@@ -496,6 +498,40 @@ describe('RecrasBooking', () => {
             // Mocks
             rb.getSetting = mockGetSetting;
             rb.getPackages = () => {
+                const packages = [{
+                    id: 26,
+                    onlineboeking_contactformulier_id: 4,
+                    regels: [
+                        {
+                            id: 669,
+                            max: 8,
+                            onlineboeking_aantalbepalingsmethode: "invullen_door_gebruiker",
+                            product: {
+                                minimum_aantal: 5,
+                                vereist_product: [],
+                            }
+                        },
+                        {
+                            id: 1337,
+                            max: null,
+                            onlineboeking_aantalbepalingsmethode: "invullen_door_gebruiker",
+                            product: {
+                                minimum_aantal: 1,
+                                vereist_product: [],
+                            }
+                        },
+                        {
+                            id: 420,
+                            aantal_personen: 10,
+                            max: 80,
+                            onlineboeking_aantalbepalingsmethode: "invullen_door_gebruiker",
+                            product: {
+                                minimum_aantal: 1,
+                                vereist_product: [],
+                            }
+                        },
+                    ],
+                }];
                 rb.packages = packages;
                 return packages;
             };